Replaced ExternalProject based dependency management with Conan, also fixed type casting errors when building for x86

+++ modules/tsgraphics/src/TextureImporter.h
@@ -31,7 +31,7 @@ namespace ts
 	{
 	private:
 
-		uint64 m_token = 0;
+		uintptr m_token = 0;
 		uint32 m_state = 0;
 
 		std::vector<uint32> m_buffer;
